requirements.txt
dhash
discord.py >= 1.4.1
emoji
GitPython
hjson
Pillow >= 7.0
pre-commit
psycopg2-binary
requests
sqlalchemy
